pyproject.toml
[tool.poetry]
name = "repl_nix_Wall-a-bot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.2.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
